+++ controllers/ingressUpdater.go
@@ -4,19 +4,23 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/cloudogu/cesapp-lib/registry"
+	"github.com/cloudogu/k8s-service-discovery/controllers/dogustart"
 	corev1 "k8s.io/api/core/v1"
 	networking "k8s.io/api/networking/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
 const (
-	staticContentBackendName       = "nginx-static"
-	staticContentBackendPort       = 80
-	staticContentBackendRewrite    = "/errors/503.html"
-	ingressRewriteTargetAnnotation = "nginx.ingress.kubernetes.io/rewrite-target"
+	staticContentBackendName           = "nginx-static"
+	staticContentBackendPort           = 80
+	staticContentBackendRewrite        = "/errors/503.html"
+	staticContentDoguIsStartingRewrite = "/errors/starting.html"
+	ingressRewriteTargetAnnotation     = "nginx.ingress.kubernetes.io/rewrite-target"
 )
 
 // CesService contains information about one exposed ces service.
@@ -34,57 +38,157 @@ type CesService struct {
 type ingressUpdater struct {
 	// client used to communicate with k8s.
 	client client.Client
+	// registry is used to read from the etcd.
+	registry registry.Registry
 	// Namespace defines the target namespace for the ingress objects.
 	namespace string
 	// IngressClassName defines the ingress class for the ces services.
 	ingressClassName string
+	// deploymentReadyChecker checks whether dogu are ready (healthy).
+	deploymentReadyChecker DeploymentReadyChecker
+}
+
+// DeploymentReadyChecker checks the readiness from deployments.
+type DeploymentReadyChecker interface {
+	// IsReady checks whether the application of the deployment is ready, i.e., contains at least one ready pod.
+	IsReady(ctx context.Context, deploymentName string) (bool, error)
 }
 
 // NewIngressUpdater creates a new instance responsible for updating ingress objects.
-func NewIngressUpdater(client client.Client, namespace string, ingressClassName string) *ingressUpdater {
-	return &ingressUpdater{
-		client:           client,
-		namespace:        namespace,
-		ingressClassName: ingressClassName,
+func NewIngressUpdater(client client.Client, registry registry.Registry, namespace string, ingressClassName string) (*ingressUpdater, error) {
+	restConfig, err := ctrl.GetConfig()
+	if err != nil {
+		return nil, fmt.Errorf("failed to find cluster config: %w", err)
 	}
+
+	clientSet, err := kubernetes.NewForConfig(restConfig)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create client set: %w", err)
+	}
+
+	deploymentReadyChecker := dogustart.NewDeploymentReadyChecker(clientSet, namespace)
+	return &ingressUpdater{
+		client:                 client,
+		registry:               registry,
+		namespace:              namespace,
+		ingressClassName:       ingressClassName,
+		deploymentReadyChecker: deploymentReadyChecker,
+	}, nil
 }
 
-// UpdateIngressOfService creates or updates the ingress object of the given service.
-func (i *ingressUpdater) UpdateIngressOfService(ctx context.Context, service *corev1.Service, isMaintenanceMode bool) error {
-	logger := log.FromContext(ctx)
-
-	if len(service.Spec.Ports) <= 0 {
-		logger.Info(fmt.Sprintf("service [%s] has no ports -> skipping ingress creation", service.Name))
-		return nil
-	}
-
-	cesServicesAnnotation, ok := service.Annotations[CesServiceAnnotation]
-	if !ok {
-		logger.Info(fmt.Sprintf("found no [%s] annotation for [%s] -> creating no ingress resource", CesServiceAnnotation, service.Name))
-		return nil
-	}
-
-	var cesServices []CesService
-	err := json.Unmarshal([]byte(cesServicesAnnotation), &cesServices)
+// UpsertIngressForService creates or updates the ingress object of the given service.
+func (i *ingressUpdater) UpsertIngressForService(ctx context.Context, service *corev1.Service) error {
+	isMaintenanceMode, err := isMaintenanceModeActive(i.registry)
 	if err != nil {
-		return fmt.Errorf("failed to unmarshal ces services: %w", err)
+		return err
+	}
+
+	cesServices, ok, err := i.getCesServices(service)
+	if err != nil {
+		return fmt.Errorf("failed to get ces services: %w", err)
+	}
+
+	if !ok {
+		log.FromContext(ctx).Info(fmt.Sprintf("service [%s] has no ports or ces services -> skipping ingress creation", service.Name))
+		return nil
 	}
 
 	for _, cesService := range cesServices {
-		err := i.createCesServiceIngress(ctx, cesService, service, isMaintenanceMode)
+		err := i.upsertIngressForCesService(ctx, cesService, service, isMaintenanceMode)
 		if err != nil {
-			return err
+			return fmt.Errorf("failed to create ingress object for ces service [%+v]: %w", cesService, err)
 		}
 	}
 
 	return nil
 }
 
-// createCesServiceIngress creates a new ingress resource based on the given ces service.
-func (i *ingressUpdater) createCesServiceIngress(ctx context.Context, cesService CesService, service *corev1.Service, isMaintenanceMode bool) error {
-	logger := log.FromContext(ctx)
-	logger.Info(fmt.Sprintf("create ces service ingress object for service [%s]", service.GetName()))
+func (i *ingressUpdater) getCesServices(service *corev1.Service) ([]CesService, bool, error) {
+	if len(service.Spec.Ports) <= 0 {
+		return []CesService{}, false, nil
+	}
 
+	cesServicesAnnotation, ok := service.Annotations[CesServiceAnnotation]
+	if !ok {
+		return []CesService{}, false, nil
+	}
+
+	var cesServices []CesService
+	err := json.Unmarshal([]byte(cesServicesAnnotation), &cesServices)
+	if err != nil {
+		return []CesService{}, false, fmt.Errorf("failed to unmarshal ces services: %w", err)
+	}
+
+	return cesServices, true, nil
+}
+
+func (i *ingressUpdater) upsertIngressForCesService(ctx context.Context, cesService CesService, service *corev1.Service, isMaintenanceMode bool) error {
+	if isMaintenanceMode {
+		return i.upsertMaintenanceModeIngressObject(ctx, cesService, service)
+	}
+
+	if hasDoguLabel(service) {
+		isReady, err := i.deploymentReadyChecker.IsReady(ctx, service.Name)
+		if err != nil {
+			return err
+		}
+
+		if !isReady {
+			return i.upsertDoguIsStartingIngressObject(ctx, cesService, service)
+		}
+	}
+
+	return i.upsertDoguIngressObject(ctx, cesService, service)
+}
+
+func (i *ingressUpdater) upsertMaintenanceModeIngressObject(ctx context.Context, cesService CesService, service *corev1.Service) error {
+	log.FromContext(ctx).Info(fmt.Sprintf("system is in maintenance mode -> create maintenance ingress object for service [%s]", service.GetName()))
+	annotations := map[string]string{ingressRewriteTargetAnnotation: staticContentBackendRewrite}
+
+	err := i.upsertIngressObject(ctx, service, cesService, staticContentBackendName, staticContentBackendPort, annotations)
+	if err != nil {
+		return fmt.Errorf("failed to update ingress object: %w", err)
+	}
+
+	return nil
+}
+
+func (i *ingressUpdater) upsertDoguIsStartingIngressObject(ctx context.Context, cesService CesService, service *corev1.Service) error {
+	log.FromContext(ctx).Info(fmt.Sprintf("dogu is still starting -> create dogu is starting ingress object for service [%s]", service.GetName()))
+	annotations := map[string]string{ingressRewriteTargetAnnotation: staticContentDoguIsStartingRewrite}
+
+	err := i.upsertIngressObject(ctx, service, cesService, staticContentBackendName, staticContentBackendPort, annotations)
+	if err != nil {
+		return fmt.Errorf("failed to update ingress object: %w", err)
+	}
+
+	return nil
+}
+
+func (i *ingressUpdater) upsertDoguIngressObject(ctx context.Context, cesService CesService, service *corev1.Service) error {
+	log.FromContext(ctx).Info(fmt.Sprintf("dogu is ready -> update ces service ingress object for service [%s]", service.GetName()))
+	annotations := map[string]string{}
+
+	if cesService.Pass != cesService.Location {
+		annotations[ingressRewriteTargetAnnotation] = cesService.Pass
+	}
+
+	err := i.upsertIngressObject(ctx, service, cesService, service.GetName(), int32(cesService.Port), annotations)
+	if err != nil {
+		return fmt.Errorf("failed to update ingress object: %w", err)
+	}
+
+	return nil
+}
+
+func (i *ingressUpdater) upsertIngressObject(
+	ctx context.Context,
+	service *corev1.Service,
+	cesService CesService,
+	endpointName string,
+	endpointPort int32,
+	annotations map[string]string,
+) error {
 	pathType := networking.PathTypePrefix
 	ingress := &networking.Ingress{
 		ObjectMeta: v1.ObjectMeta{
@@ -95,20 +199,7 @@ func (i *ingressUpdater) createCesServiceIngress(ctx context.Context, cesService
 	}
 
 	_, err := ctrl.CreateOrUpdate(ctx, i.client, ingress, func() error {
-		ingress.Annotations = map[string]string{}
-
-		serviceName := service.GetName()
-		servicePort := int32(cesService.Port)
-
-		if cesService.Pass != cesService.Location {
-			ingress.Annotations[ingressRewriteTargetAnnotation] = cesService.Pass
-		}
-
-		if isMaintenanceMode && serviceName != staticContentBackendName {
-			serviceName = staticContentBackendName
-			servicePort = staticContentBackendPort
-			ingress.Annotations[ingressRewriteTargetAnnotation] = staticContentBackendRewrite
-		}
+		ingress.Annotations = annotations
 
 		ingress.Spec = networking.IngressSpec{
 			IngressClassName: &i.ingressClassName,
@@ -119,9 +210,9 @@ func (i *ingressUpdater) createCesServiceIngress(ctx context.Context, cesService
 							PathType: &pathType,
 							Backend: networking.IngressBackend{
 								Service: &networking.IngressServiceBackend{
-									Name: serviceName,
+									Name: endpointName,
 									Port: networking.ServiceBackendPort{
-										Number: servicePort,
+										Number: endpointPort,
 									},
 								}}}}}}}}}
 
@@ -136,4 +227,15 @@ func (i *ingressUpdater) createCesServiceIngress(ctx context.Context, cesService
 	}
 
 	return nil
+}
+
+func isMaintenanceModeActive(r registry.Registry) (bool, error) {
+	_, err := r.GlobalConfig().Get(maintenanceModeGlobalKey)
+	if registry.IsKeyNotFoundError(err) {
+		return false, nil
+	} else if err != nil {
+		return false, fmt.Errorf("failed to read the maintenance mode from the registry: %w", err)
+	}
+
+	return true, nil
 }