+++ controllers/serviceController_int_test.go
@@ -7,6 +7,11 @@ import (
 	"context"
 	_ "embed"
 	"fmt"
+	"github.com/cloudogu/k8s-service-discovery/controllers/dogustart"
+	"k8s.io/api/apps/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	ctrl "sigs.k8s.io/controller-runtime"
 	"time"
 
 	"k8s.io/apimachinery/pkg/types"
@@ -35,7 +40,9 @@ var serviceWebApp = &corev1.Service{}
 var serviceAdditionalBytes []byte
 var serviceAdditional = &corev1.Service{}
 
-func init() {
+var serviceDeployment = &v1.Deployment{}
+
+func resetData() {
 	err := yaml.Unmarshal(serviceNoAnnotationsBytes, serviceNoAnnotations)
 	if err != nil {
 		panic(err)
@@ -50,10 +57,38 @@ func init() {
 	if err != nil {
 		panic(err)
 	}
+
+	serviceDeployment = &v1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "nexus",
+			Namespace: "my-test-namespace",
+			Labels:    map[string]string{"dogu.name": "nexus"},
+		},
+		Spec: v1.DeploymentSpec{
+			Selector: &metav1.LabelSelector{
+				MatchLabels: map[string]string{"dogu.name": "nexus"},
+			},
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels: map[string]string{"dogu.name": "nexus"},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{Name: "nexus-123123", Image: "nginx:1.7.9"},
+					},
+				},
+			},
+		},
+		Status: v1.DeploymentStatus{
+			Replicas:      0,
+			ReadyReplicas: 0,
+		},
+	}
 }
 
 var _ = Describe("Creating ingress objects with the ingress generator", func() {
-	ctx := context.TODO()
+	ctx := context.Background()
+	resetData()
 
 	Context("Handle new service resource", func() {
 		AfterEach(cleanup)
@@ -74,12 +109,55 @@ var _ = Describe("Creating ingress objects with the ingress generator", func() {
 		})
 
 		It("Should create ingress object for simple webapp service", func() {
+			By("Create deployment for service (which is not ready)")
+			Expect(k8sClient.Create(ctx, serviceDeployment)).Should(Succeed())
+
 			By("Creating service with ces annotations")
 			Expect(k8sClient.Create(ctx, serviceWebApp)).Should(Succeed())
 
-			By("Expect exactly one ingress resource for the service")
+			By("Expect exactly one ingress resource for the dogu is starting ingress object")
 			expectedIngress := &networking.IngressList{}
 
+			Eventually(func() bool {
+				err := k8sClient.List(ctx, expectedIngress)
+				if err != nil {
+					_ = fmt.Errorf("%w", err)
+					return false
+				}
+
+				return len(expectedIngress.Items) == 1
+			}, timeoutInterval, pollingInterval).Should(BeTrue())
+
+			Expect(expectedIngress.Items[0].Namespace).Should(Equal(myNamespace))
+			Expect(expectedIngress.Items[0].Name).Should(Equal("nexus"))
+			Expect(*expectedIngress.Items[0].Spec.IngressClassName).Should(Equal(myIngressClassName))
+			Expect(expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].Path).Should(Equal("/nexus"))
+			Expect(*expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].PathType).Should(Equal(networking.PathTypePrefix))
+			Expect(expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].Backend.Service.Name).Should(Equal("nginx-static"))
+			Expect(expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].Backend.Service.Port.Number).Should(Equal(int32(80)))
+
+			value, ok := expectedIngress.Items[0].Annotations[ingressRewriteTargetAnnotation]
+			Expect(ok).Should(BeTrue())
+			Expect(value).Should(Equal("/errors/starting.html"))
+
+			By("Wait for deployment to become ready")
+			serviceDeployment.Status.ReadyReplicas = 1
+			serviceDeployment.Status.Replicas = 1
+			err := k8sClient.Status().Update(ctx, serviceDeployment)
+			Expect(err).NotTo(HaveOccurred())
+
+			client, err := kubernetes.NewForConfig(ctrl.GetConfigOrDie())
+			Expect(err).NotTo(HaveOccurred())
+			deploymentWaiter := dogustart.NewDeploymentReadyChecker(client, "my-test-namespace")
+			waitOptions := dogustart.WaitOptions{Timeout: time.Minute * 30, TickRate: time.Millisecond * 200}
+			err = deploymentWaiter.WaitForReady(ctx, "nexus", waitOptions, func(ctx context.Context) {})
+			Expect(err).NotTo(HaveOccurred())
+
+			By("Wait until async onReady method was called")
+			time.Sleep(time.Second * 2)
+
+			By("Expect exactly one ingress resource for the service")
+			expectedIngress = &networking.IngressList{}
 			Eventually(func() bool {
 				err := k8sClient.List(ctx, expectedIngress)
 				if err != nil {
@@ -98,11 +176,27 @@ var _ = Describe("Creating ingress objects with the ingress generator", func() {
 			Expect(expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].Backend.Service.Name).Should(Equal("nexus"))
 			Expect(expectedIngress.Items[0].Spec.Rules[0].HTTP.Paths[0].Backend.Service.Port.Number).Should(Equal(int32(8082)))
 
-			_, ok := expectedIngress.Items[0].Annotations[ingressRewriteTargetAnnotation]
+			_, ok = expectedIngress.Items[0].Annotations[ingressRewriteTargetAnnotation]
 			Expect(ok).Should(BeFalse())
 		})
 
 		It("Should create ingress object for multiple ces services", func() {
+			By("Create deployment for service (which is not ready)")
+			Expect(k8sClient.Create(ctx, serviceDeployment)).Should(Succeed())
+
+			By("Wait for deployment to become ready")
+			serviceDeployment.Status.ReadyReplicas = 1
+			serviceDeployment.Status.Replicas = 1
+			err := k8sClient.Status().Update(ctx, serviceDeployment)
+			Expect(err).NotTo(HaveOccurred())
+
+			client, err := kubernetes.NewForConfig(ctrl.GetConfigOrDie())
+			Expect(err).NotTo(HaveOccurred())
+			deploymentWaiter := dogustart.NewDeploymentReadyChecker(client, "my-test-namespace")
+			waitOptions := dogustart.WaitOptions{Timeout: time.Minute * 30, TickRate: time.Millisecond * 200}
+			err = deploymentWaiter.WaitForReady(ctx, "nexus", waitOptions, func(ctx context.Context) {})
+			Expect(err).NotTo(HaveOccurred())
+
 			By("Creating service with multiple ces services")
 			Expect(k8sClient.Create(ctx, serviceAdditional)).Should(Succeed())
 
@@ -166,7 +260,8 @@ var _ = Describe("Ingress class should be created automatically", func() {
 })
 
 func cleanup() {
-	ctx := context.TODO()
+	ctx := context.Background()
+	resetData()
 
 	By("Cleanup all ingresses")
 	ingressesList := &networking.IngressList{}
@@ -196,6 +291,24 @@ func cleanup() {
 
 		for _, service := range servicesList.Items {
 			err = k8sClient.Delete(ctx, &service)
+			if err != nil {
+				return false
+			}
+		}
+
+		return true
+	}, timeoutInterval, pollingInterval).Should(BeTrue())
+
+	By("Cleanup all deployments")
+	deploymentList := &v1.DeploymentList{}
+	Eventually(func() bool {
+		err := k8sClient.List(ctx, deploymentList)
+		if err != nil {
+			return false
+		}
+
+		for _, deployment := range deploymentList.Items {
+			err = k8sClient.Delete(ctx, &deployment)
 			if err != nil {
 				return false
 			}